renamed AutoFrameworkLinear and EncodeAuto
swapped names so that they correctly corresponded to the classes' actual content
