+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/EncoderAuto.java
@@ -29,11 +29,13 @@
 
 package org.firstinspires.ftc.teamcode;
 
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
+import com.qualcomm.robotcore.util.Range;
+
 
 
 /**
@@ -52,6 +54,10 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 @TeleOp(name="Basic: Linear OpMode", group="Linear Opmode")
 @Disabled
 public class EncoderAuto extends LinearOpMode {
+
+    private double TICKS_PER_ROTATION=1500;
+    private double INCHES_PER_ROTATION=3; //wheel's circumference
+    private double TICKS_PER_INCH=TICKS_PER_ROTATION*INCHES_PER_ROTATION;
 
     // Declare OpMode members.
     private ElapsedTime runtime = new ElapsedTime();
@@ -75,10 +81,14 @@ public class EncoderAuto extends LinearOpMode {
         backLeftDrive = hardwareMap.get(DcMotor.class, "back_left_drive");
         backRightDrive = hardwareMap.get(DcMotor.class, "back_right_drive");
 
+        all_run_using_encoders();
+
         // Most robots need the motor on one side to be reversed to drive forward
         // Reverse the motor that runs backwards when connected directly to the battery
         backLeftDrive.setDirection(DcMotor.Direction.REVERSE);
         frontLeftDrive.setDirection(DcMotor.Direction.REVERSE);
+
+
 
         // Wait for the game to start (driver presses PLAY)
         waitForStart();
@@ -88,15 +98,16 @@ public class EncoderAuto extends LinearOpMode {
         while (opModeIsActive()) {
             if (stage == 0) {
                 mecanum(1, 0, 0);
-                if (runtime.seconds()>=5) {
-                    runtime.reset();
+                if (Math.abs(frontLeftDrive.getCurrentPosition())>=10*TICKS_PER_INCH){
                     stage++;
+                    reset_encoders();
                 }
                 //robot moves forward for 5 seconds
             } else if (stage==1) {
-                mecanum(0, 1, 0);
-                if (runtime.seconds()>=5) {
+                mecanum(0, -1, 0);
+                if (frontLeftDrive.getCurrentPosition()>=10*TICKS_PER_INCH) {
                     stage++;
+                    reset_encoders();
                 }
                 //robot moves to the side for 5 seconds
             }
@@ -132,6 +143,22 @@ public class EncoderAuto extends LinearOpMode {
         backLeftDrive.setPower(backLeftPower);
         frontRightDrive.setPower(frontRightPower);
         backRightDrive.setPower(backRightPower);
+    }
+
+    public void reset_encoders() {
+        frontLeftDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backLeftDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        backRightDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        frontRightDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+
+        all_run_using_encoders();
+    }
+
+    public void all_run_using_encoders() {
+        backLeftDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        backRightDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        frontRightDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        frontLeftDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
     }
